Update react-native loader to match browser updates
