+++ lib/react-native-loader.js
@@ -1,12 +1,6 @@
 var AWS = require('./core');
 // react-native specific modules
-AWS.util.crypto.lib = {
-  createHash: require('create-hash'),
-  createHmac: require('create-hmac'),
-  createSign: function() {
-    throw new Error('createSign is not implemented in the browser');
-  }
-}
+AWS.util.crypto.lib = require('./browserCryptoLib');
 AWS.util.Buffer = require('buffer/').Buffer;
 AWS.util.url = require('url/');
 AWS.util.querystring = require('querystring/');
